feat(30956513) : added better preJoin ui
